PROTOCOL CHANGE: Reads are now assumed to be data reads
Reading the device configuration requires an i2c "Write" command
to tell the device to return its configuration information.

Right now, configuration is all a no-op. And in day to day operation,
configuartion is requested much, much less rarely than key data updates.

The new wire protocol assumes that a read without a prior write is asking
for key data. This has the nice effect of halving the number of messages over
the i2c bus.